Crate iup_sys

Rust low level bindings for IUP -- A Portable User Interface Toolkit

These bindings follow what is found in iup.h as closely as possibly with the following exceptions:

C function-like macros have been converted into functions.

Items that were marked as old or deprecated have been removed.

Enums

Ihandle

Constants

IUP_BOTTOM

IUP_BUTTON1

IUP_BUTTON2

IUP_BUTTON3

IUP_BUTTON4

IUP_BUTTON5

IUP_CENTER

IUP_CENTERPARENT

IUP_CLOSE

IUP_CONTINUE

IUP_COPYRIGHT

IUP_CURRENT

IUP_DEFAULT

IUP_DESCRIPTION

IUP_ERROR

IUP_GETPARAM_CANCEL

IUP_GETPARAM_HELP

IUP_GETPARAM_INIT

IUP_GETPARAM_OK

IUP_HIDE

IUP_IGNORE

IUP_INVALID

IUP_INVALID_ID

IUP_LEFT

IUP_MASK_EFLOAT

IUP_MASK_FLOAT

IUP_MASK_INT

IUP_MASK_UFLOAT

IUP_MASK_UINT

IUP_MAXIMIZE

IUP_MINIMIZE

IUP_MOUSEPOS

IUP_NAME

IUP_NOERROR

IUP_OPENED

IUP_RECBINARY

IUP_RECTEXT

IUP_RESTORE

IUP_RIGHT

IUP_SBDN

IUP_SBDRAGH

IUP_SBDRAGV

IUP_SBLEFT

IUP_SBPGDN

IUP_SBPGLEFT

IUP_SBPGRIGHT

IUP_SBPGUP

IUP_SBPOSH

IUP_SBPOSV

IUP_SBRIGHT

IUP_SBUP

IUP_SHOW

IUP_TOP

IUP_VERSION

IUP_VERSION_DATE

IUP_VERSION_NUMBER

Functions

IupAlarm

IupAppend

IupBackgroundBox

IupButton

IupCanvas

IupCbox

IupCboxv

IupClassMatch

IupClipboard

IupClose

IupColorDlg

IupConvertXYToPos

IupCopyClassAttributes

IupCreate

IupCreatep

IupCreatev

IupDestroy

IupDetach

IupDetachBox

IupDialog

IupElementPropertiesDialog

IupExitLoop

IupExpander

IupFileDlg

IupFill

IupFlush

IupFontDlg

IupFrame

IupGetAllAttributes

IupGetAllClasses

IupGetAllDialogs

IupGetAllNames

IupGetAttribute

IupGetAttributeHandle

IupGetAttributeId

IupGetAttributeId2

IupGetAttributes

IupGetBrother

IupGetCallback

IupGetChild

IupGetChildCount

IupGetChildPos

IupGetClassAttributes

IupGetClassCallbacks

IupGetClassName

IupGetClassType

IupGetColor

IupGetDialog

IupGetDialogChild

IupGetDouble

IupGetDoubleId

IupGetDoubleId2

IupGetFile

IupGetFloat

IupGetFloatId

IupGetFloatId2

IupGetFocus

IupGetFunction

IupGetGlobal

IupGetHandle

IupGetInt

IupGetInt2

IupGetIntId

IupGetIntId2

IupGetIntInt

IupGetLanguage

IupGetLanguageString

IupGetName

IupGetNextChild

IupGetParam

IupGetParamv

IupGetParent

IupGetRGB

IupGetRGBId

IupGetRGBId2

IupGetText

IupGridBox

IupGridBoxv

IupHbox

IupHboxv

IupHelp

IupHide

IupImage

IupImageLibOpen

IupImageRGB

IupImageRGBA

IupInsert

IupItem

IupLabel

IupLayoutDialog

IupLink

IupList

IupListDialog

IupLoad

IupLoadBuffer

IupLoopStep

IupLoopStepWait

IupMainLoop

IupMainLoopLevel

IupMap

IupMenu

IupMenuv

IupMessage

IupMessageDlg

IupMessagef

IupMultiLine

IupNextField

IupNormalizer

IupNormalizerv

IupOpen

IupPlayInput

IupPopup

IupPreviousField

IupProgressBar

IupProgressDlg

IupRadio

IupRecordInput

IupRedraw

IupRefresh

IupRefreshChildren

IupReparent

IupResetAttribute

IupSaveClassAttributes

IupSaveImageAsText

IupSbox

IupScanf

IupScrollBox

IupSeparator

IupSetAtt

IupSetAttribute

IupSetAttributeHandle

IupSetAttributeId

IupSetAttributeId2

IupSetAttributes

IupSetCallback

IupSetCallbacks

IupSetClassDefaultAttribute

IupSetDouble

IupSetDoubleId

IupSetDoubleId2

IupSetFloat

IupSetFloatId

IupSetFloatId2

IupSetFocus

IupSetFunction

IupSetGlobal

IupSetHandle

IupSetInt

IupSetIntId

IupSetIntId2

IupSetLanguage

IupSetLanguagePack

IupSetLanguageString

IupSetRGB

IupSetRGBId

IupSetRGBId2

IupSetStrAttribute

IupSetStrAttributeId

IupSetStrAttributeId2

IupSetStrGlobal

IupSetStrf

IupSetStrfId

IupSetStrfId2

IupShow

IupShowXY

IupSplit

IupStoreLanguageString

IupSubmenu

IupTabs

IupTabsv

IupText

IupTextConvertLinColToPos

IupTextConvertPosToLinCol

IupTimer

IupToggle

IupTree

IupTreeGetId

IupTreeGetUserId

IupTreeSetAttributeHandle

IupTreeSetUserId

IupUnmap

IupUpdate

IupUpdateChildren

IupUser

IupVal

IupVbox

IupVboxv

IupVersion

IupVersionDate

IupVersionNumber

IupZbox

IupZboxv

iup_isalt

iup_isbutton1

iup_isbutton2

iup_isbutton3

iup_isbutton4

iup_isbutton5

iup_iscontrol

iup_isdouble

iup_isshift

iup_issys

Type Definitions

Icallback

Iparamcb